backend/requirements.txt
fastapi
uvicorn
python-dotenv
websockets
openai
pydantic
